+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/infrastructure/edc/notificationcontract/service/EdcNotificationContractServiceTest.java
@@ -21,12 +21,13 @@
 
 package org.eclipse.tractusx.traceability.infrastructure.edc.notificationcontract.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import org.eclipse.tractusx.irs.edc.client.asset.EdcAssetService;
 import org.eclipse.tractusx.irs.edc.client.asset.model.exception.CreateEdcAssetException;
+import org.eclipse.tractusx.irs.edc.client.asset.model.exception.DeleteEdcAssetException;
 import org.eclipse.tractusx.irs.edc.client.contract.model.exception.CreateEdcContractDefinitionException;
 import org.eclipse.tractusx.irs.edc.client.contract.service.EdcContractDefinitionService;
 import org.eclipse.tractusx.irs.edc.client.policy.model.exception.CreateEdcPolicyDefinitionException;
+import org.eclipse.tractusx.irs.edc.client.policy.model.exception.DeleteEdcPolicyDefinitionException;
 import org.eclipse.tractusx.irs.edc.client.policy.service.EdcPolicyDefinitionService;
 import org.eclipse.tractusx.traceability.common.properties.TraceabilityProperties;
 import org.eclipse.tractusx.traceability.qualitynotification.application.contract.model.CreateNotificationContractRequest;
@@ -39,10 +40,12 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.web.client.RestTemplate;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -51,10 +54,7 @@ class EdcNotificationContractServiceTest {
     @Mock
     TraceabilityProperties traceabilityProperties;
     @Mock
-    EdcAssetService edcAssetService;
-
-    @Mock
-    RestTemplate restTemplate;
+    EdcAssetService edcNotificationAssetService;
 
     @Mock
     EdcPolicyDefinitionService edcPolicyDefinitionService;
@@ -63,33 +63,91 @@ class EdcNotificationContractServiceTest {
     EdcContractDefinitionService edcContractDefinitionService;
 
     private EdcNotificationContractService edcNotificationContractService;
-    private CreateNotificationContractRequest request;
+
 
     private static final String notificationAssetId = "9";
     private static final String accessPolicyId = "99";
     private static final String contractDefinitionId = "999";
 
     @BeforeEach
-    void setUp() throws JsonProcessingException, CreateEdcAssetException, CreateEdcPolicyDefinitionException, CreateEdcContractDefinitionException {
-        String rightOperand = "trace3";
-        NotificationType notificationType = NotificationType.QUALITY_INVESTIGATION;
-        NotificationMethod notificationMethod = NotificationMethod.RESOLVE;
-        request = new CreateNotificationContractRequest(notificationType, notificationMethod);
-        when(edcAssetService.createNotificationAsset(any(), any(), any(), any())).thenReturn(notificationAssetId);
-        when(traceabilityProperties.getRightOperand()).thenReturn(rightOperand);
-//        when(edcNotificationAssetService.createNotificationAsset(notificationMethod, request.notificationType())).thenReturn(notificationAssetId);
-        when(edcPolicyDefinitionService.createAccessPolicy(rightOperand)).thenReturn(accessPolicyId);
-        when(edcContractDefinitionService.createContractDefinition(notificationAssetId, accessPolicyId)).thenReturn(contractDefinitionId);
+    void setUp() {
         edcNotificationContractService = new EdcNotificationContractService(
-                edcAssetService, edcPolicyDefinitionService, edcContractDefinitionService, traceabilityProperties
+                edcNotificationAssetService, edcPolicyDefinitionService, edcContractDefinitionService, traceabilityProperties
         );
     }
 
     @Test
-    void testHandle() {
+    void testHandle() throws CreateEdcAssetException, CreateEdcPolicyDefinitionException, CreateEdcContractDefinitionException {
+        // given
+        String rightOperand = "trace3";
+        NotificationType notificationType = NotificationType.QUALITY_INVESTIGATION;
+        NotificationMethod notificationMethod = NotificationMethod.RESOLVE;
+        CreateNotificationContractRequest request = new CreateNotificationContractRequest(notificationType, notificationMethod);
+        when(edcNotificationAssetService.createNotificationAsset(any(), any(), any(), any())).thenReturn(notificationAssetId);
+        when(traceabilityProperties.getUrl()).thenReturn("https://test");
+        when(traceabilityProperties.getRightOperand()).thenReturn(rightOperand);
+        when(edcPolicyDefinitionService.createAccessPolicy(rightOperand)).thenReturn(accessPolicyId);
+        when(edcContractDefinitionService.createContractDefinition(notificationAssetId, accessPolicyId)).thenReturn(contractDefinitionId);
+
+        // when
         CreateNotificationContractResponse response = edcNotificationContractService.handle(request);
+
+        // then
         assertThat(notificationAssetId).isEqualTo(response.notificationAssetId());
         assertThat(accessPolicyId).isEqualTo(response.accessPolicyId());
         assertThat(contractDefinitionId).isEqualTo(response.contractDefinitionId());
+        verify(edcNotificationAssetService).createNotificationAsset(
+                "https://test/api/qualitynotifications/resolve",
+                "QUALITY_INVESTIGATION RESOLVE",
+                org.eclipse.tractusx.irs.edc.client.asset.model.NotificationMethod.RESOLVE,
+                org.eclipse.tractusx.irs.edc.client.asset.model.NotificationType.QUALITY_INVESTIGATION);
+    }
+
+    @Test
+    void givenService_whenAssetCreationThrowsException_thenThrowException() throws CreateEdcAssetException, CreateEdcPolicyDefinitionException, CreateEdcContractDefinitionException {
+        // given
+        NotificationType notificationType = NotificationType.QUALITY_INVESTIGATION;
+        NotificationMethod notificationMethod = NotificationMethod.RESOLVE;
+        CreateNotificationContractRequest request = new CreateNotificationContractRequest(notificationType, notificationMethod);
+        when(traceabilityProperties.getUrl()).thenReturn("https://test");
+        doThrow(CreateEdcAssetException.class).when(edcNotificationAssetService).createNotificationAsset(any(), any(), any(), any());
+
+        // when/then
+        assertThrows(RuntimeException.class, () -> edcNotificationContractService.handle(request));
+    }
+
+    @Test
+    void givenService_whenPolicyDefinitionServiceThrowsException_thenThrowException() throws CreateEdcAssetException, CreateEdcPolicyDefinitionException, DeleteEdcAssetException {
+        // given
+        String rightOperand = "trace3";
+        NotificationType notificationType = NotificationType.QUALITY_INVESTIGATION;
+        NotificationMethod notificationMethod = NotificationMethod.RESOLVE;
+        CreateNotificationContractRequest request = new CreateNotificationContractRequest(notificationType, notificationMethod);
+        when(edcNotificationAssetService.createNotificationAsset(any(), any(), any(), any())).thenReturn(notificationAssetId);
+        when(traceabilityProperties.getUrl()).thenReturn("https://test");
+        when(traceabilityProperties.getRightOperand()).thenReturn(rightOperand);
+        doThrow(CreateEdcPolicyDefinitionException.class).when(edcPolicyDefinitionService).createAccessPolicy(any());
+
+        // when/then
+        assertThrows(RuntimeException.class, () -> edcNotificationContractService.handle(request));
+        verify(edcNotificationAssetService).deleteAsset(any());
+    }
+
+    @Test
+    void givenService_whenContractDefinitionServiceThrowsException_thenThrowException() throws CreateEdcAssetException, CreateEdcContractDefinitionException, DeleteEdcAssetException, DeleteEdcPolicyDefinitionException {
+        // given
+        String rightOperand = "trace3";
+        NotificationType notificationType = NotificationType.QUALITY_INVESTIGATION;
+        NotificationMethod notificationMethod = NotificationMethod.RESOLVE;
+        CreateNotificationContractRequest request = new CreateNotificationContractRequest(notificationType, notificationMethod);
+        when(edcNotificationAssetService.createNotificationAsset(any(), any(), any(), any())).thenReturn(notificationAssetId);
+        when(traceabilityProperties.getUrl()).thenReturn("https://test");
+        when(traceabilityProperties.getRightOperand()).thenReturn(rightOperand);
+        doThrow(CreateEdcContractDefinitionException.class).when(edcContractDefinitionService).createContractDefinition(any(), any());
+
+        // when/then
+        assertThrows(RuntimeException.class, () -> edcNotificationContractService.handle(request));
+        verify(edcPolicyDefinitionService).deleteAccessPolicy(any());
+        verify(edcNotificationAssetService).deleteAsset(any());
     }
 }