+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 ### Removed
+- removed EDC notification asset classes and replaced with IRS lib implementation
 
 ## [10.5.0 - 22.02.2024]